requirements.txt
uvicorn
fastapi[standard]
snowflake-snowpark-python
pandas
google-genai
dotenv
sqlalchemy
snowflake-sqlalchemy